+++ src/instructions/mod.rs
@@ -4,16 +4,10 @@
 
 pub mod interrupts;
 pub mod port;
+pub mod random;
 pub mod segmentation;
 pub mod tables;
 pub mod tlb;
-
-/// Cause a breakpoint exception by invoking the `int3` instruction.
-pub fn int3() {
-    unsafe {
-        asm!("int3" :::: "volatile");
-    }
-}
 
 /// Halts the CPU until the next interrupt arrives.
 #[inline(always)]