+++ src/lib.rs
@@ -4,7 +4,6 @@
 #![feature(const_fn)]
 #![feature(asm)]
 #![feature(abi_x86_interrupt)]
-#![feature(try_from)]
 #![cfg_attr(not(test), no_std)]
 #![cfg_attr(feature = "deny-warnings", deny(warnings))]
 #![cfg_attr(feature = "deny-warnings", deny(missing_docs))]
@@ -19,7 +18,6 @@ pub use ux;
 pub use crate::addr::{align_down, align_up, PhysAddr, VirtAddr};
 
 pub mod instructions;
-pub mod random;
 pub mod registers;
 pub mod structures;